split MRMeshDistance
