+++ source/MRMesh/MRMesh.vcxproj.filters
@@ -1230,6 +1230,9 @@
     <ClInclude Include="MRScopedValue.h">
       <Filter>Source Files\Basic</Filter>
     </ClInclude>
+    <ClInclude Include="MRMeshMeshDistance.h">
+      <Filter>Source Files\AABBTree</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="MRObject.cpp">
@@ -2020,6 +2023,9 @@
     </ClCompile>
     <ClCompile Include="MRIOFilters.cpp">
       <Filter>Source Files\IO</Filter>
+    </ClCompile>
+    <ClCompile Include="MRMeshMeshDistance.cpp">
+      <Filter>Source Files\AABBTree</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
